# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/roles/{id}": {
      "get": {
        "description": "只能查自己账号下的角色（否则 403）。\n\n---\n\n**鉴权**：客户端令牌（`player`），只能操作令牌所属账号自己的数据。",
        "operationId": "role.detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "角色完整数据",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "player": []
          }
        ],
        "summary": "角色详情",
        "tags": [
          "客户端"
        ],
        "x-olua-audience": "player",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/RoleDto"
        }
      },
      "put": {
        "description": "全量覆盖 data；id 以路径为准（请求体里的 id 会被忽略）。\n\n请求体可带 `revision`（客户端上次读到的修订号）：带上即开启**乐观锁**，与服务端当前值不一致时返回 20006 （角色已在别处被修改），客户端应先拉一次详情、以最新数据为基线再继续推；不带则不做冲突校验。\n\n---\n\n**鉴权**：客户端令牌（`player`），只能操作令牌所属账号自己的数据。",
        "operationId": "role.save",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存后的角色",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "player": []
          }
        ],
        "summary": "保存角色进度",
        "tags": [
          "客户端"
        ],
        "x-olua-audience": "player",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/RoleDto"
        }
      },
      "delete": {
        "description": "不可恢复；删的正好是在线角色时同时清掉在线标记。\n\n---\n\n**鉴权**：客户端令牌（`player`），只能操作令牌所属账号自己的数据。",
        "operationId": "role.remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功（data 为 null）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true,
                          "additionalProperties": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "player": []
          }
        ],
        "summary": "删除角色",
        "tags": [
          "客户端"
        ],
        "x-olua-audience": "player",
        "x-olua-data-nullable": true
      }
    },
    "/roles/{id}/select": {
      "post": {
        "description": "把该角色设为账号的在线角色。\n\n---\n\n**鉴权**：客户端令牌（`player`），只能操作令牌所属账号自己的数据。",
        "operationId": "role.select",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "被选中的角色",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "player": []
          }
        ],
        "summary": "选中（进入）角色",
        "tags": [
          "客户端"
        ],
        "x-olua-audience": "player",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/RoleDto"
        }
      }
    },
...
    "/admin/roles": {
      "get": {
        "description": "keyword 匹配角色名或角色 id；accountId 限定账号；online / occupation / sex / minLevel / maxLevel 做筛选。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`role:read`（查看角色；角色：超级管理员 / 管理员 / 只读观察员）",
        "operationId": "adminRole.list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "每页条数（1~100）",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字（不同接口含义见其说明）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "description": "限定账号 id（查某账号下的全部角色）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "online",
            "required": false,
            "in": "query",
            "description": "在线状态筛选：true 只看在线（账号当前选中的）角色",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "occupation",
            "required": false,
            "in": "query",
            "description": "职业筛选",
            "schema": {
              "type": "string",
              "enum": [
                "1",
                "2",
                "3",
                "4"
              ]
            }
          },
          {
            "name": "sex",
            "required": false,
            "in": "query",
            "description": "性别筛选",
            "schema": {
              "type": "string",
              "enum": [
                "1",
                "2",
                "3"
              ]
            }
          },
          {
            "name": "minLevel",
            "required": false,
            "in": "query",
            "description": "最低等级（含）",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "type": "number"
            }
          },
          {
            "name": "maxLevel",
            "required": false,
            "in": "query",
            "description": "最高等级（含）",
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页结果（list/total/page/size）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminRolePageDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "角色列表",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminRolePageDto"
        },
        "x-olua-permissions": [
          "role:read"
        ],
        "x-olua-query-schema": {
          "$ref": "#/components/schemas/RoleQueryDto"
        }
      }
    },
    "/admin/roles/batch-delete": {
      "post": {
        "description": "按 id 列表删除（最多 100 条）。已不存在的 id 静默跳过（幂等）；作为某个账号在线角色的会被顺带清掉在线标记。返回实际删除的条数与涉及的账号 id。\n\nPOST 而不是 DELETE 是因为要带请求体；用 200 而不是 201 —— 它不创建资源。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`role:delete`（删除角色；角色：超级管理员 / 管理员）",
        "operationId": "adminRole.batchRemove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除结果（requested / deleted / ids / clearedOnlineAccountIds）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BatchDeleteResultDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "批量删除角色",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/BatchDeleteResultDto"
        },
        "x-olua-permissions": [
          "role:delete"
        ]
      }
    },
    "/admin/roles/{id}": {
      "get": {
        "description": "含完整 data（背包 / 装备等原样返回）。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`role:read`（查看角色；角色：超级管理员 / 管理员 / 只读观察员）",
        "operationId": "adminRole.detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "角色详情",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminRoleDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "角色详情",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminRoleDto"
        },
        "x-olua-permissions": [
          "role:read"
        ]
      },
      "patch": {
        "description": "只提交要改的字段，其余原样保留。开放三类：**基础信息**（角色名 / 职业 / 性别 / 等级 / 时装 / 头像 / 所在地图）、**常用数值**（金币 / 绑定元宝 / 银两 / 经验 / 战魂 / 称号 / 军衔）、**运行时数据**（装备穿戴表 / 技能等级表 / 背包格子，结构化编辑，传 `bag: []` 即清空背包）。\n\n角色名仍需满足「同一账号下不重名」。保存后修订号 +1，在线玩家的下一次进度推送会先同步到这里。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`role:write`（修改角色；角色：超级管理员 / 管理员）",
        "operationId": "adminRole.patch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPatchRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改后的角色",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminRoleDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "修改角色",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminRoleDto"
        },
        "x-olua-permissions": [
          "role:write"
        ]
      },
      "delete": {
        "description": "不可恢复；是账号在线角色时一并清掉在线标记。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`role:delete`（删除角色；角色：超级管理员 / 管理员）",
        "operationId": "adminRole.remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功（data 为 null）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true,
                          "additionalProperties": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "删除角色",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": true,
        "x-olua-permissions": [
          "role:delete"
        ]
      }
    },
    "/admin/roles/{id}/select": {
      "post": {
        "description": "把角色设为所属账号的在线角色（等价于玩家在选角界面选它进游戏）。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`role:select`（切换在线角色；角色：超级管理员 / 管理员）",
        "operationId": "adminRole.select",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "被设为在线的角色",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminRoleDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "设为在线角色",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminRoleDto"
        },
        "x-olua-permissions": [
          "role:select"
        ]
      }
    },
...
    "/admin/admins/{id}": {
      "patch": {
        "description": "改角色或启停。允许改自己（超管轮值），但**最后一个启用中的超级管理员**不能被降级 / 停用 —— 避免没人能进后台。改完立即生效。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`admin:manage`（管理管理员；角色：超级管理员）",
        "operationId": "adminAdmin.update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "管理员 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改后的管理员",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "修改管理员",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": false,
        "x-olua-data-schema": {
          "$ref": "#/components/schemas/AdminDto"
        },
        "x-olua-permissions": [
          "admin:manage"
        ]
      },
      "delete": {
        "description": "不可恢复；同样保护最后一个启用中的超级管理员。\n\n---\n\n**鉴权**：管理端令牌（`admin`）。\n\n**所需权限**：`admin:manage`（管理管理员；角色：超级管理员）",
        "operationId": "adminAdmin.remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "管理员 id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功（data 为 null）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiEnvelopeDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "nullable": true,
                          "additionalProperties": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "admin": []
          }
        ],
        "summary": "删除管理员",
        "tags": [
          "管理端"
        ],
        "x-olua-audience": "admin",
        "x-olua-data-nullable": true,
        "x-olua-permissions": [
          "admin:manage"
        ]
      }
    },
...
  },
  "info": {
    "title": "olua 游戏服务端 API",
    "description": "olua 客户端与管理端共用的服务端接口。\n\n**接口分三组**（左侧按组展示，顺序即权限层级）：\n1. **公共接口** —— 无需令牌（注册 / 登录 / 健康检查）\n2. **客户端** —— 需要 `player` 令牌（`/api/auth/login` 获取），只能操作自己账号的数据\n3. **管理端** —— 需要 `admin` 令牌（`/api/admin/auth/login` 获取），且每个接口还要求权限点\n\n**统一响应包裹**：所有接口（成功与失败）都是 `{ code, message, data, timestamp }`；\n`code = 0` 表示成功，其余见服务端 `common/constants/biz-code.ts`。\n\n**权限**：管理员分三级 —— 超级管理员 / 管理员 / 只读观察员，\n每个管理端接口的说明里都写了「所需权限」，由 `common/constants/permission.ts` 统一维护。\n请求头一律是 `Authorization: Bearer <token>`。\n\n---\n\n**这份文档是三端接口文件的唯一来源**：客户端 `assets/ui/utils/net/` 与管理端 `admin/src/api/`\n里的路径、类型、接口方法都由 `tools/gen-api.cjs` 从本文件生成，\n所以**改接口先改服务端**，再执行 `cd server && npm run gen:api` 重新生成（禁止手改生成物）。\n生成器读取的机器可读约定：\n- 每个接口的 `operationId` 形如 `<模块>.<方法>`（如 `role.save`），决定生成到哪个模块的哪个方法；\n- `x-olua-audience` 标明调用方（public / player / admin）；\n- `x-olua-permissions` 是管理端接口要求的权限点；\n- `x-olua-data-schema` 是响应里 `data` 的真实类型（没有该扩展 = 该接口无业务数据）。",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "player": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "客户端令牌（/api/auth/login 获取）"
      },
      "admin": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "管理端令牌（/api/admin/auth/login 获取）"
      }
    },
    "schemas": {
...
      "SaveRoleDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "description": "角色完整数据（全量覆盖）",
            "additionalProperties": true
          },
          "revision": {
            "type": "number",
            "description": "客户端读到的修订号（带上即开启乐观锁：与服务端不一致时返回 20006，需先拉最新数据）",
            "example": 3
          }
        },
        "required": [
          "data"
        ]
      },
...
      "BatchDeleteRolesDto": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "要删除的角色 id 列表（最多 100 条）",
            "example": [
              "1760000000000"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
...
      "ApiEnvelopeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "业务码，0 = 成功",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "ok"
          },
          "data": {
            "type": "object",
            "description": "业务数据（无数据时为 null）",
            "additionalProperties": true,
            "nullable": true
          },
          "timestamp": {
            "type": "number",
            "description": "服务端时间戳（毫秒）",
            "example": 1760000000000
          },
          "path": {
            "type": "string",
            "description": "出错时的请求路径",
            "example": "/api/auth/login"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "timestamp"
        ]
      },
...
      "RoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "角色 id"
          },
          "accountId": {
            "type": "string",
            "description": "所属账号 id"
          },
          "name": {
            "type": "string",
            "description": "角色名"
          },
          "occupation": {
            "type": "string",
            "description": "职业：1 战士 / 2 魔法师 / 3 道士 / 4 全职业"
          },
          "sex": {
            "type": "string",
            "description": "性别：1 男 / 2 女 / 3 全性别"
          },
          "level": {
            "type": "number",
            "description": "等级"
          },
          "online": {
            "type": "boolean",
            "description": "是否当前选中（在线）角色"
          },
          "revision": {
            "type": "number",
            "description": "修订号（每次落库 +1；客户端保存进度时带上它做乐观锁）"
          },
          "createdAt": {
            "type": "number",
            "description": "创建时间（毫秒）"
          },
          "updatedAt": {
            "type": "number",
            "description": "最近更新时间（毫秒）"
          },
          "data": {
            "type": "object",
            "description": "角色完整数据（客户端 entities/Role 的 JSON 快照，服务端按不透明文档存取；管理端只编辑其中的常用字段）",
            "additionalProperties": true
          }
        },
        "required": [
          "id",
          "accountId",
          "name",
          "occupation",
          "sex",
          "level",
          "online",
          "revision",
          "createdAt",
          "updatedAt",
          "data"
        ]
      },
...
      "AdminRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "角色 id"
          },
          "accountId": {
            "type": "string",
            "description": "所属账号 id"
          },
          "name": {
            "type": "string",
            "description": "角色名"
          },
          "occupation": {
            "type": "string",
            "description": "职业：1 战士 / 2 魔法师 / 3 道士 / 4 全职业"
          },
          "sex": {
            "type": "string",
            "description": "性别：1 男 / 2 女 / 3 全性别"
          },
          "level": {
            "type": "number",
            "description": "等级"
          },
          "online": {
            "type": "boolean",
            "description": "是否当前选中（在线）角色"
          },
          "revision": {
            "type": "number",
            "description": "修订号（每次落库 +1；客户端保存进度时带上它做乐观锁）"
          },
          "createdAt": {
            "type": "number",
            "description": "创建时间（毫秒）"
          },
          "updatedAt": {
            "type": "number",
            "description": "最近更新时间（毫秒）"
          },
          "data": {
            "type": "object",
            "description": "角色完整数据（客户端 entities/Role 的 JSON 快照，服务端按不透明文档存取；管理端只编辑其中的常用字段）",
            "additionalProperties": true
          },
          "accountName": {
            "type": "string",
            "description": "所属账号名",
            "nullable": true
          }
        },
        "required": [
          "id",
          "accountId",
          "name",
          "occupation",
          "sex",
          "level",
          "online",
          "revision",
          "createdAt",
          "updatedAt",
          "data",
          "accountName"
        ]
      },
      "AdminRolePageDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "总条数",
            "example": 42
          },
          "page": {
            "type": "number",
            "description": "当前页码（从 1 开始）",
            "example": 1
          },
          "size": {
            "type": "number",
            "description": "每页条数",
            "example": 20
          },
          "list": {
            "description": "当前页的角色列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminRoleDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "size",
          "list"
        ]
      },
      "BatchDeleteResultDto": {
        "type": "object",
        "properties": {
          "requested": {
            "type": "number",
            "description": "请求删除的条数"
          },
          "deleted": {
            "type": "number",
            "description": "实际删除的条数（已不存在的 id 不计入）"
          },
          "ids": {
            "description": "实际被删掉的角色 id",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "clearedOnlineAccountIds": {
            "description": "被顺带清掉在线角色标记的账号 id（删的是在线角色时）",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "requested",
          "deleted",
          "ids",
          "clearedOnlineAccountIds"
        ]
      },
...
      "RoleQueryDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码，从 1 开始",
            "example": 1,
            "default": 1,
            "minimum": 1
          },
          "size": {
            "type": "number",
            "description": "每页条数（1~100）",
            "example": 20,
            "default": 20,
            "minimum": 1,
            "maximum": 100
          },
          "keyword": {
            "type": "string",
            "description": "关键字（不同接口含义见其说明）"
          },
          "accountId": {
            "type": "string",
            "description": "限定账号 id（查某账号下的全部角色）"
          },
          "online": {
            "type": "string",
            "description": "在线状态筛选：true 只看在线（账号当前选中的）角色",
            "enum": [
              "true",
              "false"
            ]
          },
          "occupation": {
            "type": "string",
            "description": "职业筛选",
            "enum": [
              "1",
              "2",
              "3",
              "4"
            ]
          },
          "sex": {
            "type": "string",
            "description": "性别筛选",
            "enum": [
              "1",
              "2",
              "3"
            ]
          },
          "minLevel": {
            "type": "number",
            "description": "最低等级（含）",
            "minimum": 1,
            "maximum": 1000
          },
          "maxLevel": {
            "type": "number",
            "description": "最高等级（含）",
            "minimum": 1,
            "maximum": 1000
          }
        }
      },
...
    }
  }
}